LinkedIn_AIAgent/requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
requests-html>=0.10.0
python-dotenv>=1.0.0
spacy>=3.7.0
lxml>=5.0.0

# After install, also run:
# python -m spacy download en_core_web_sm
